feat(data): drive sampler/test validation through TaskQuerySchema
Addresses the Copilot + @mmcdermott review on #96 calling for the code to
take more advantage of the schema class.  Changes:

* **Schema-native dtypes.**  ``TaskQuerySchema`` keeps ``query`` as
  ``pa.large_string`` (polars' ``Utf8`` serialises to ``large_string``
  when a DataFrame is converted to arrow; verified via ``df.to_arrow()``)
  and ``duration_days`` as ``pa.float32``.  Sampler's ``build_index_df``
  now emits ``pl.Float32`` directly instead of ``pl.Int64``, so producer
  output aligns to the schema natively without an ``align()``-time
  coercion step.

* **Override ``LabelSchema.boolean_value`` to allow nulls.**  MEDS's
  ``LabelSchema`` ships ``boolean_value`` as ``Optional(bool, nullable=NONE)``;
  EveryQuery's convention (closes #122) uses null as the "censored"
  sentinel, so we override to ``Optional(bool, nullable=True)``.  Keeps
  the column optional (inference-only inputs without labels still
  validate) but allows the three-valued task-label semantics.

* **Sampler ``evaluate_index_df`` empty path uses ``TaskQuerySchema``**
  instead of a hand-rolled polars-schema dict: build an empty arrow
  table from ``TaskQuerySchema.schema()`` and project to the 5-column
  label shape.

* **Sampler ``run_worker`` validates against the schema at the write
  boundary.**  ``TaskQuerySchema.validate(labeled.to_arrow())`` fires
  before ``_atomic_write_parquet``, so any producer drift (wrong dtype,
  missing column, extra column) fails loudly at write time rather than
  silently corrupting downstream pipelines.

* **Evaluate ``gen_task.py`` emits ``Float32`` durations** (was
  implicit ``int32`` from ``pl.lit(duration)``) so its output is also
  schema-conformant.

* **Tests use ``TaskQuerySchema.validate()`` for schema assertions**
  instead of hand-rolled column-set checks:
  - ``tests/test_eval_suite.py::TestGenTaskOutputSchema::test_columns``
    → ``test_conforms_to_task_query_schema``; dropped the redundant
    ``test_non_censored_rows_are_non_null`` which was tautological under
    the collapsed schema.
  - ``tests/test_sample_tasks.py::TestRunWorkerPipeline`` column-set
    asserts and ``TestEndToEndWithDataset`` column-set assert both
    replaced with ``TaskQuerySchema.validate(...)``.
  - Test index_dfs that feed ``evaluate_index_df`` cast
    ``duration_days`` to ``Float32`` to match the schema contract.

All 205 non-slow tests pass locally in 118s.

On the separate question about renaming the post-processed ``boolean_value``
column in ``dataset.py`` schema_df to ``censored``: kept as ``boolean_value``
for super-class compatibility — ``MEDSPytorchDataset.LABEL_COL`` hard-codes
``"boolean_value"`` and wires it through parquet read, ``self.labels``
caching, and the per-item dict key used in ``collate``.  Overriding
``LABEL_COL = "censored"`` would either lose super's parquet-read path
(which keys off ``LABEL_COL`` being present in the parquet schema names)
or require subclassing considerably more of it.  The clean name exists
at the batch-API layer: ``EveryQueryBatch.censor`` is the explicit
censor-indicator tensor downstream code consumes; ``batch.boolean_value``
(super-inherited, same tensor) is effectively an alias.

Refs #80, #96, #122.

+++ src/every_query/evaluate/gen_task.py
@@ -99,7 +99,9 @@ def process_eval_tasks(
                     .with_columns(
                         boolean_value.alias("boolean_value"),
                         pl.lit(code).alias("query"),
-                        pl.lit(duration).alias("duration_days"),
+                        # Float32 to match ``TaskQuerySchema.duration_days`` so the
+                        # output aligns to the schema natively.
+                        pl.lit(duration).cast(pl.Float32).alias("duration_days"),
                     )
                     .select(
                         "subject_id",
